src/main/java/com/ap/qa/pages/HomePage.java
package com.ap.qa.pages;

import org.openqa.selenium.WebElement;
import org.openqa.selenium.support.FindBy;
import org.openqa.selenium.support.PageFactory;

import com.ap.qa.base.TestBase;


public class HomePage extends TestBase{

	//object repository for home page
	@FindBy(xpath = "//*[@id=\"block_top_menu\"]/ul/li[3]/a")
	WebElement tshirtLink;


	// Initializing the Page Objects
	public HomePage() {
		PageFactory.initElements(driver, this);
	}
	
	public String verifyHomePageTitle(){
		return driver.getTitle();
	}
	
	
	public AddToCartPage clickOnTshirtLink(){
		tshirtLink.click();
		return new AddToCartPage();
	}
}
